finish survey answer logic

+++ src/main/java/com/midgetspinner31/survey/db/entity/Restrictions.java
@@ -13,8 +13,8 @@ import java.util.List;
 @AllArgsConstructor
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class Restrictions {
-    Double min;
-    Double max;
+    Integer min;
+    Integer max;
     Integer maxLength;
     List<String> choices;
